backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
websockets==15.0.1
sqlmodel==0.0.33
python-dotenv==1.2.2
requests==2.32.5
httpx==0.28.1
beautifulsoup4==4.14.3
alembic==1.18.4
bcrypt==5.0.0
PyJWT==2.10.1
python-multipart==0.0.22
